Add descriptor
