+++ packages/types/src/index.ts
@@ -19,7 +19,6 @@ export { default as BlockNumber } from './BlockNumber';
 export { default as Bool } from './Bool';
 export { default as Bytes } from './Bytes';
 export { default as Extrinsic } from './Extrinsic';
-// NOTE Only used internally
 export { default as ExtrinsicIndex } from './ExtrinsicIndex';
 // NOTE Only used internally, exported as PendingExtrinsics
 // export { default as Extrinsics } from './Extrinsics';
@@ -62,7 +61,6 @@ export { default as u32 } from './U32';
 export { default as u64 } from './U64';
 export { default as u128 } from './U128';
 export { default as u256 } from './U256';
-export { default as UncheckedExtrinsic } from './UncheckedExtrinsic';
 export { default as ValidatorPrefs } from './ValidatorPrefs';
 export { default as VoteThreshold } from './VoteThreshold';
 export { default as VoteIndex } from './VoteIndex';
